fix: types to own file
